default values validate as valid, begin adding configuration for number of days
