--- src/renderer/configuration/ConfigurationFormProps.tsx
@@ -1,8 +1,0 @@
-import { FormType } from "../Form"
-
-type ConfigurationFormProps = {
-  id: string
-  onSubmit: (evt: React.SyntheticEvent, form: FormType) => void
-}
-
-export default ConfigurationFormProps
